move report generation logic into generator package, #131
