+++ generator/e2e_test.go
@@ -80,3 +80,40 @@ func TestEndToEndHTMLGeneration(t *testing.T) {
 		helper.AssertEqual(want, got, expectedFile, t)
 	}
 }
+
+func TestEndToEndHTMLGenerationFromSavedResult(t *testing.T) {
+	expectedFiles := []string{"index.html", "passing_specification_1.html", "failing_specification_1.html", "skipped_specification.html", "js/search_index.js"}
+	reportDir := filepath.Join("_testdata", "e2e")
+	inputFile := filepath.Join("_testdata", "last_run_result.json")
+
+	RegenerateReport(inputFile, reportDir, templateBasePath)
+	for _, expectedFile := range expectedFiles {
+		gotContent, err := ioutil.ReadFile(filepath.Join(reportDir, expectedFile))
+		if err != nil {
+			t.Errorf("Error reading generated HTML file: %s", err.Error())
+		}
+		wantContent, err := ioutil.ReadFile(filepath.Join("_testdata", "expectedE2E", expectedFile))
+		if err != nil {
+			t.Errorf("Error reading expected HTML file: %s", err.Error())
+		}
+		got := helper.RemoveNewline(string(gotContent))
+		want := helper.RemoveNewline(string(wantContent))
+		os.Remove(filepath.Join(reportDir, expectedFile))
+		helper.AssertEqual(want, got, expectedFile, t)
+	}
+	cleanUp(t)
+}
+
+func cleanUp(t *testing.T) {
+	reportDir := filepath.Join("_testdata", "e2e")
+	os.RemoveAll(filepath.Join(reportDir, "css"))
+	os.RemoveAll(filepath.Join(reportDir, "images"))
+	os.RemoveAll(filepath.Join(reportDir, "fonts"))
+	s, err := filepath.Glob(filepath.Join(reportDir, "js", "*.js"))
+	if err != nil {
+		t.Error(err)
+	}
+	for _, f := range s {
+		os.Remove(f)
+	}
+}
